widened test coverage and moved under a new file for better doctest module behaviour
